unsigned longs changed to uint32_t

+++ typeClassDefs/VisionGoals.h
@@ -39,7 +39,7 @@ private:
 	wb_goal bottomLeft;//bottom camera
     wb_goal bottomRight;
     wb_goal bottomGeneric;
-	unsigned long _frameNumber;
+	uint32_t _frameNumber;
 public:
 	/**Default Constructor*/
     VisionGoals() : topLeft(), topRight(), bottomLeft(), bottomRight(), _frameNumber(0) {
@@ -275,7 +275,7 @@ public:
 	 * @brief Sets the frame number this information in this message was observed.
      * @param fn The frame number
      */
-	void setFrameNumber(unsigned long fn) {
+	void setFrameNumber(uint32_t fn) {
 		_frameNumber = fn;
 	}
 	
@@ -283,7 +283,7 @@ public:
 	 * @brief Get the frame number the information in this message was observed.
      * @return The frame number
      */
-	unsigned long frameNumber() const {
+	uint32_t frameNumber() const {
 		return _frameNumber;
 	}
